api/requirements.txt
fastapi
uvicorn[standard]
pydantic[email]
python-multipart
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
email-validator
stripe
python-socketio
